// src/shared/enums.ts
export enum BUTTON_VARIANT_ENUM {
  PRIMARY = "primary",
  PRIMARY_GRADIENT = "primary-gradient",
  WHITE = "white",
  SECONDARY = "secondary",
  TERTIARY = "tertiary",
  TERTIARY_PRIMARY = "tertiary-primary",
  TERTIARY_WHITE = "tertiary-white",
}

export enum BUTTON_SIZE_ENUM {
  DEFAULT = "default",
  SMALL = "small",
  LARGE = "large",
}

export enum BUTTON_BORDER_RADIUS_ENUM {
  DEFAULT = "default",
  SMALL = "small",
  FULL = "full",
}